weather/serializers.py
from rest_framework import serializers

from .models import WeatherRecord


class WeatherRecordSerializer(serializers.ModelSerializer):
    class Meta:
        model = WeatherRecord
        fields = ['id', 'city', 'temperature', 'humidity', 'timestamp']
